Noticed an issue with the previous sanitization setup.
inputs do not need htmlspecialchars, as we are using prepared statements in the repositories. any injected code will not affect the database or the backend.

When using placeholders (e.g., :username, :email, :password) in the SQL query and then bind the actual values to these placeholders, we prevent any part of the input from being interpreted as SQL code. Instead, everything that the user inputs is treated purely as data, even if it contains SQL-like syntax.

for script injections (XSS), those run on a local browser level. This means that using htmlspecialchars on output is sufficient enough to prevent XSS.

Issue fixed:
htmlspecialchars were a part of sanitize function. this meant that users would not be able to use characters like " ' &
htmlspecialchars would then filter and convert those to predefined html entities which would then be stored in the database.
fx.
"What's up"
would get stored in the database as:
"what&#039;s up"

In short, Prepared statements (in our case for example bindparam) in the repositories along with trim function is enough to prevent injections. and htmlspecialchars on output will prevent XSS.