adjustments/dependencies
Basecontroller Added

The basecontroller serves as a base for common dependencies used in controllers.

Authservice added.
Some logic related to authentication has moved to authservice.

Those logic were previously implemented in authcontroller. Which meant that controllers had to instantiate new userservice and new authcontroller to access them. OR to rely on static functions.

In order to mainting project standards, the logic is now called from authservice.